v0.1.6 add aggregates
